pyproject.toml
[tool.poetry]
name = "audiobook_tools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
mutagen = "^1.46.0"
pydub = "^0.25.1"
python-dateutil = "^2.8.2"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
